Added FockBackend base class and implemented compatible backend attribute to all tasks
