+++ lightworks/emulator/backends/__init__.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .abc_backend import BackendABC
 from .backend import Backend
+from .fock_backend import FockBackend
 from .permanent import PermanentBackend
 from .slos import SLOSBackend
